Debug GUI: 3-page console with gnome controls, item spawning, and game settings
Gnomes page: gnome selector dropdown (refreshes on window open), hunger/thirst/sleep
sliders with set buttons, kill/spawn creature buttons, window size selector.

Items page: cascading group/item/material dropdowns replacing text fields. Composite
items get a second material dropdown. Temporary component items are destroyed after
composite creation.

Game page: need decay multiplier slider with live text display, per-need decay disable
checkboxes.

Also adds Global::debugNeedDecayMultiplier and Global::disabledNeedDecays for gnome
need decay control from the debug console.

+++ src/game/gnome.h
@@ -52,7 +52,7 @@ public:
 	//return true if need needs action
 	bool evalNeeds( bool seasonChanged, bool dayChanged, bool hourChanged, bool minuteChanged );
 
-	void addNeed( QString id, int level );
+	void addNeed( QString id, float level );
 	int need( QString id );
 
 	void selectProfession( QString profession );
